+++ src/openclatura/assembly_parts.py
@@ -2,6 +2,7 @@
 
 from dataclasses import dataclass, field
 
+from .locant_sources import LocantMapSource
 from .name_operations import HydroOperation
 from .spiro_assembly import SpiroAssembly
 
@@ -156,7 +157,7 @@ class AssemblyParts:
     parent_atom_charges_by_locant: dict[str, int] = field(default_factory=dict)
     parent_bond_orders_by_locants: dict[tuple[str, str], int] = field(default_factory=dict)
     parent_bond_ids_by_locants: dict[tuple[str, str], int] = field(default_factory=dict)
-    locant_map_source: str = "generated"
+    locant_map_source: LocantMapSource = LocantMapSource.GENERATED
     name_atom_bindings: list[NameAtomBinding] = field(default_factory=list)
     name_token_spans: list[dict] = field(default_factory=list)
     name_rewrite_history: list[dict] = field(default_factory=list)
