+++ src/openclatura/locant_sources.py
@@ -1,0 +1,11 @@
+"""Provenance labels for parent locant maps."""
+
+from enum import StrEnum
+
+
+class LocantMapSource(StrEnum):
+    """Where the parent atom-to-locant map came from."""
+
+    GENERATED = "generated"
+    SUPPLIED = "supplied"
+    PROOF = "proof"
